requirements.txt
jupyterlab
matplotlib
numpy
pandas
pynrrd
pyradiomics
scikit-learn
seaborn